+++ charts/core/src/Tooltip/CustomTooltip/SeriesList/SeriesList.types.ts
@@ -1,8 +1,8 @@
 import { ReactNode } from 'react';
 
-import { AxisFormatterCallbackParams } from '../Tooltip.types';
+import { AxisFormatterCallbackParams } from '../../Tooltip.types';
 
-export interface TooltipSeriesListProps {
+export interface SeriesListProps {
   params: AxisFormatterCallbackParams;
   seriesValueFormatter?: (value: string | number | Date) => string | ReactNode;
   seriesNameFormatter?: (seriesName: string) => string | ReactNode;
