+++ charts/core/src/Tooltip/CustomTooltip/SeriesList/index.ts
@@ -1,0 +1,2 @@
+export { SeriesList } from './SeriesList';
+export type { SeriesListProps } from './SeriesList.types';
