+++ charts/core/src/Tooltip/CustomTooltip/SeriesListItem/index.ts
@@ -1,0 +1,2 @@
+export { SeriesListItem } from './SeriesListItem';
+export type { SeriesListItemProps } from './SeriesListItem.types';
