+++ charts/core/src/Tooltip/CustomTooltip/index.ts
@@ -1,0 +1,1 @@
+export { CustomTooltip } from './CustomTooltip';
