--- charts/core/src/Tooltip/TooltipContent/index.ts
@@ -1,1 +1,0 @@
-export { TooltipContent } from './TooltipContent';
