--- charts/core/src/Tooltip/TooltipSeriesList/index.ts
@@ -1,2 +1,0 @@
-export { TooltipSeriesList } from './TooltipSeriesList';
-export type { TooltipSeriesListProps } from './TooltipSeriesList.types';
