--- charts/core/src/Tooltip/TooltipSeriesListItem/index.ts
@@ -1,2 +1,0 @@
-export { TooltipSeriesListItem } from './TooltipSeriesListItem';
-export type { TooltipSeriesListItemProps } from './TooltipSeriesListItem.types';
